travel_ai_backend/app/requirements.txt
firebase_admin==6.5.0
Flask==3.0.3
Flask_Cors==4.0.1
langchain_openai==0.2.2
protobuf
python-dotenv==1.0.1
google-cloud-secret-manager==2.20.2
gunicorn==23.0.0
google-generativeai~=0.7.2
